refactor: separate code
