+++ .github/workflows/auto-approval.yml
@@ -4,8 +4,8 @@ jobs:
   auto-approve:
     runs-on: ubuntu-latest
     steps:
-    - name: Approve Dependabot updates
-      uses: hmarr/auto-approve-action@v2.0.0
-      if: github.actor == 'dependabot[bot]' || github.actor == 'dependabot-preview[bot]'
-      with:
-        github-token: "${{ secrets.USE_GITHUB_TOKEN }}"
+      - name: Approve Dependabot updates
+        uses: hmarr/auto-approve-action@v2.0.0
+        if: github.actor == 'dependabot[bot]' || github.actor == 'dependabot-preview[bot]'
+        with:
+          github-token: "${{ secrets.USE_GITHUB_TOKEN }}"
