+++ src/core/get_info.ts
@@ -1,0 +1,18 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import { Context } from "probot";
+/* eslint-enable  @typescript-eslint/no-unused-vars*/
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import type { UserInfo } from "../utils/types";
+/* eslint-enable  @typescript-eslint/no-unused-vars*/
+
+/**
+ * Parses the user information from the event payload.
+ * 
+ * @param context The Probot context.
+ */
+export const getUserInfo = (context: Context): UserInfo => {
+  const info: UserInfo = {
+    username: context.payload.pull_request.user.login,
+  };
+  return info;
+};
