+++ src/core/index.ts
@@ -1,190 +1,17 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import type { ReviewEvent, ReviewLookupResult, UserInfo } from "../utils/types";
-/* eslint-enable  @typescript-eslint/no-unused-vars*/
-import {
-  composeCrashReportDetails,
-  composeCrashReportSummary,
-  composeCrashReportTitle,
-  composeReviewDismissalMsg,
-  composeStatusCheckDetails,
-  composeStatusCheckSummary,
-  composeStatusCheckTitle,
-} from "../utils/msg_composer";
 import { containsNotAllowedFile, ownsAllFiles } from "../utils/rule_matcher";
+import {
+  createCrashStatus,
+  createPassingStatus,
+  createStartStatus,
+} from "./check_status";
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { Context } from "probot";
 /* eslint-enable  @typescript-eslint/no-unused-vars*/
-import { StatusCodes } from "http-status-codes";
-import { createStatus } from "./create_status";
-import { getAppActorName } from "../utils/config";
+import { approveChange } from "./review";
+import { dismissAllApprovals } from "./dismiss_approval";
+import { getChangedFiles } from "./get_files";
 import { getOwnershipRules } from "../utils/config_parser";
-
-const getPullAuthor = (context: Context): string => {
-  return context.payload.pull_request.user.login;
-};
-
-const getUserInfo = (context: Context): UserInfo => {
-  const info: UserInfo = {
-    username: getPullAuthor(context),
-  };
-  return info;
-};
-
-const approveChange = async (context: Context): Promise<void> => {
-  const req = context.repo({
-    "event": "APPROVE" as ReviewEvent,
-    "pull_number": context.payload.pull_request.number,
-  });
-  context.log.info(`Reviewing PR with request ${JSON.stringify(req)}`);
-  try {
-    const res = await context.github.pulls.createReview(req);
-    if (res.status === StatusCodes.OK) {
-      context.log.info("Approve changes succeeded.");
-    } else {
-      context.log.error(
-        `Approve change rejected with: ${JSON.stringify(res.data)}`,
-      );
-    }
-  } catch (err) {
-    context.log.error(`Approve change failed with: ${JSON.stringify(err)}`);
-  }
-};
-
-/**
- * Create a check in the pull request to indicate that the app
- * has received and processed the pull request event.
- *
- * Note:
- *
- * - It can only be passing or not exist at all.
- * - It is for indicating that the event is handled.
- * - It indicates nothing about whether the PR is approved.
- *
- * @param context The request context created by Probot core.
- * @param startTime The timestamp that this run started.
- */
-export const createPassingStatus = async (
-  context: Context,
-  startTime: string,
-): Promise<void> => {
-  await createStatus(
-    context,
-    "success",
-    "completed",
-    composeStatusCheckTitle(),
-    composeStatusCheckSummary(),
-    composeStatusCheckDetails(),
-    startTime,
-  );
-};
-
-/**
- * Create a status in pull requets indicating that the run
- * has started.
- *
- * @param context The request context from Probot core.
- * @param startTime The string time that the run started.
- */
-export const createStartStatus = async (
-  context: Context,
-  startTime: string,
-): Promise<void> => {
-  await createStatus(
-    context,
-    /* eslint-disable */
-    undefined,
-    /* eslint-enable */
-    "in_progress",
-    composeStatusCheckTitle(),
-    composeStatusCheckSummary(),
-    composeStatusCheckDetails(),
-    startTime,
-  );
-};
-
-export const createCrashStatus = async (
-  context: Context,
-  startTime: string,
-  error: Error,
-): Promise<void> => {
-  await createStatus(
-    context,
-    "failure",
-    "completed",
-    composeCrashReportTitle(),
-    composeCrashReportSummary(),
-    composeCrashReportDetails(error),
-    startTime,
-  );
-};
-
-export const getChangedFiles = async (context: Context): Promise<string[]> => {
-  const req = context.repo({
-    "pull_number": context.payload.pull_request.number,
-  });
-  const changedFilesResponse = await context.github.pulls.listFiles(req);
-  const changedFiles: string[] = [];
-  for (const changedFileData of changedFilesResponse.data) {
-    changedFiles.push(changedFileData.filename);
-  }
-  context.log.info(`Changed files are: ${JSON.stringify(changedFiles)}`);
-  return changedFiles;
-};
-
-const getPreviousReviewIds = async (
-  context: Context,
-): Promise<ReviewLookupResult> => {
-  const req = context.repo({
-    "pull_number": context.payload.pull_request.number,
-  });
-  const reviewsResponse = await context.github.pulls.listReviews(req);
-  let hasReview = false;
-  const reviewIds: number[] = [];
-  context.log.info(`Found ${reviewsResponse.data.length} reviews`);
-  reviewsResponse.data.forEach((review: Record<string, unknown>) => {
-    if ("user" in review) {
-      const user = review["user"] as Record<string, unknown>;
-      if ("login" in user) {
-        const username = user["login"] as string;
-        context.log.info(`Found review left by ${username}`);
-        if (
-          username === getAppActorName() &&
-          (review["state"] as string) !== "DISMISSED"
-        ) {
-          hasReview = true;
-          reviewIds.push(review["id"] as number);
-        }
-      }
-    }
-  });
-  return { hasReview, reviewIds };
-};
-
-const dismissApproval = async (
-  context: Context,
-  reviewId: number,
-): Promise<void> => {
-  const req = context.repo({
-    "message": composeReviewDismissalMsg(),
-    "pull_number": context.payload.pull_request.number,
-    "review_id": reviewId,
-  });
-  context.log.info("Try to dismiss the review");
-  const dismissResponse = await context.github.pulls.dismissReview(req);
-  context.log.info(
-    `Dismiss review #${reviewId} in PR #${req["pull_number"]} ` +
-      `with status ${dismissResponse.status} ` +
-      `and review state ${dismissResponse.data.state}`,
-  );
-};
-
-export const dismissAllApprovals = async (context: Context): Promise<void> => {
-  const reviewLookupResult = await getPreviousReviewIds(context);
-  for (const reviewId of reviewLookupResult.reviewIds) {
-    await dismissApproval(context, reviewId);
-  }
-  context.log.info(`Dismissed ${reviewLookupResult.reviewIds.length} reviews`);
-};
+import { getUserInfo } from "./get_info";
 
 /**
  * Approves a pull request if all the files it contains are owned by the user that
